Permanently supervise a Brain process
I don’t love calling Brain.set in test setup, though, but I guess it makes sense. Starting the Brain process with existing data felt somehow cleaner than using its public API, but as I write that it feels like the opposite should be true.
