+++ lib/elix.ex
@@ -14,6 +14,7 @@ defmodule Elix do
       # worker(Elix.Worker, [arg1, arg2, arg3]),
       worker(Elix.Robot, []),
       worker(Redix, [redis_url, [name: :redix]]),
+      worker(Elix.Brain, []),
       worker(Elix.MessageScheduler, []),
     ]
