--- lib/elix/message_scheduler/store.ex
@@ -1,5 +1,0 @@
-defmodule Elix.MessageScheduler.Store do
-  @callback all :: List.t
-  @callback add(String.t, String.t) :: any
-  @callback remove(String.t) :: any
-end
